requirements.txt
 flask
 PyPDF2
 pytesseract 
 pdf2image
 Pillow
 asyncio
 joblib